Add removing workers and sheets

+++ sheetview.py
@@ -536,10 +536,12 @@ class SheetView(QGraphicsView):
 
 
     def keyPressEvent(self, event):
-        if event.key() == Qt.Key_Delete:
-            # Delete selection except for init and loop
-            nodes = [x for x in self.scene.items() if isinstance(x, Node)]
-            for node in nodes:
-                if node.isSelected() and not node == self.initnode and not node == self.loopnode:
-                    node.delete()
-                    self.scene.removeItem(node)
+        if not event.isAccepted():
+            if event.key() == Qt.Key_Delete:
+                event.accept()
+                # Delete selection except for init and loop
+                nodes = [x for x in self.scene.items() if isinstance(x, Node)]
+                for node in nodes:
+                    if node.isSelected() and not node == self.initnode and not node == self.loopnode:
+                        node.delete()
+                        self.scene.removeItem(node)
